Make LastPassify compatible with Psych 4 and above
When trying to upgrade another project to Ruby 3, we saw this error:

```
ruby/3.1.0/psych.rb:323:in `safe_load': wrong number of arguments (given 4, expected 1) (ArgumentError)
```

We found that [Psych had changed its interface to use named arguments](https://github.com/ruby/psych/pull/487/commits/176494297f3f124467a6e3f1c9e6400ee742d663)
for `safe_load` in Psych 4. So this code looks at what the current Psych
version is to see if it's 4 or above, and then calls `safe_load` with or
without named arguments, depending on what the current version is.

We did this so that lastpassify could remain backwards compatible but
also work as we move more applications to Ruby 3.

One alternative presented was to use `unsafe_load` but that felt risky
since it's a gem used in many projects.

We had to make each of the variables `Gem::Version.new` so that we'd be
comparing `versions` and not strings. I myself got a little confused
when I saw Bundler version `2.2.23` was newer than `2.2.6`. String
comparison would make the same mistake I did, so we needed `versions`
to be able to semantically compare them.

I was surprised to find `Psych::VERSION` returned a string, not a
`version`, causing the tests to error because it couldn't compare a
version against a string. This is why both of the variables needed to
use `Gem::Version.new`.

Co-authored-by: Brittany Greaner <grea0022@umn.edu>
Co-authored-by: Davin Lagerroos <dlagerro@umn.edu>